+++ Sources/HummingbirdRouter/RouteHandler.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Hummingbird server framework project
 //
-// Copyright (c) 2023 the Hummingbird authors
+// Copyright (c) 2023-2024 the Hummingbird authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
@@ -16,14 +16,18 @@ import Hummingbird
 
 /// Protocol for route handler object.
 ///
-/// Requires a function that returns a response from a request and context
-public protocol RouteHandlerProtocol<Context>: Sendable {
+/// Requires a function that returns a HTTP response from a HTTP request and context.
+@_documentation(visibility: internal)
+public protocol _RouteHandlerProtocol<Context>: Sendable {
     associatedtype Context: HBRouterRequestContext
     func handle(_ request: HBRequest, context: Context) async throws -> HBResponse
 }
 
-/// Implementatinon of RouteHandleProtocol that uses a closure to produce a response
-public struct RouteHandlerClosure<RouteOutput: HBResponseGenerator, Context: HBRouterRequestContext>: RouteHandlerProtocol {
+/// Implementatinon of ``_RouteHandlerProtocol`` that uses a closure to produce a response.
+///
+/// This is used internally to implement `Route` when it is initialized with a closure.
+@_documentation(visibility: internal)
+public struct _RouteHandlerClosure<RouteOutput: HBResponseGenerator, Context: HBRouterRequestContext>: _RouteHandlerProtocol {
     @usableFromInline
     let closure: @Sendable (HBRequest, Context) async throws -> RouteOutput
 
@@ -33,8 +37,12 @@ public struct RouteHandlerClosure<RouteOutput: HBResponseGenerator, Context: HBR
     }
 }
 
-/// Implementatinon of RouteHandleProtocol that uses a MiddlewareStack to produce a resposne
-public struct RouteHandlerMiddleware<M0: MiddlewareProtocol>: RouteHandlerProtocol where M0.Input == HBRequest, M0.Output == HBResponse, M0.Context: HBRouterRequestContext {
+/// Implementatinon of ``_RouteHandlerProtocol`` that uses a MiddlewareStack to produce a resposne
+///
+/// This is used internally to implement `Route` when it is initialized with a middleware built
+/// from the ``RouteBuilder`` result builder.
+@_documentation(visibility: internal)
+public struct _RouteHandlerMiddleware<M0: MiddlewareProtocol>: _RouteHandlerProtocol where M0.Input == HBRequest, M0.Output == HBResponse, M0.Context: HBRouterRequestContext {
     public typealias Context = M0.Context
 
     /// Dummy function passed to middleware handle
