+++ Tests/HummingbirdRouterTests/RouterTests.swift
@@ -341,7 +341,7 @@ final class RouterTests: XCTestCase {
 
     func testPartialCapture() async throws {
         let router = HBRouterBuilder(context: HBBasicRouterRequestContext.self) {
-            Get("/files/file.{ext}/{name}.jpg") { _, context -> String in
+            Route(.get, "/files/file.{ext}/{name}.jpg") { _, context -> String in
                 XCTAssertEqual(context.parameters.count, 2)
                 let ext = try context.parameters.require("ext")
                 let name = try context.parameters.require("name")
